metadata_service/requirements.txt
flask
Pillow
Flask-SQLAlchemy
flask-migrate